Setup datamegh.api.db module
