+++ datamegh/util/db.py
@@ -54,16 +54,3 @@ def build_connstr(**args) -> str:
     connstr = CONN_STR.format(**params)
 
     return connstr
-
-
-def connect(**params) -> pyodbc.Connection:
-    """ Open connection to a Database. """
-    logger.debug(
-        "Connecting to database: {}/{}".format(
-            params.get("host"), params.get("database")
-        )
-    )
-
-    connstr = build_connstr(**params)
-
-    return pyodbc.connect(connstr)
