+++ tests/util/test_db.py
@@ -2,9 +2,7 @@
 
 import pytest
 
-from mock import patch
-
-from datamegh.util.db import connect, build_connstr, PG, MSSQL, MYSQL
+from datamegh.util.db import build_connstr, PG, MSSQL, MYSQL
 
 
 def test_build_connstr_1():
@@ -113,23 +111,3 @@ def test_build_connstr_4():
         connstr_mysql
         == "DRIVER=SOME_OTHER_DRIVER;SERVER=localhost;PORT=33066;DATABASE=test;UID=test;PWD=Test@123"
     )
-
-
-@patch("datamegh.util.db.pyodbc")
-def test_connect(m_pyodbc):
-    """
-    Test it invokes pyodbc connect
-    generating a connection string.
-    """
-
-    connection = connect(
-        client=PG,
-        port=5444,
-        driver="SOME_OTHER_DRIVER",
-        host="localhost",
-        database="test",
-        username="test",
-        password="Test@123",
-    )
-    expected_connstr = "DRIVER=SOME_OTHER_DRIVER;SERVER=localhost;PORT=5444;DATABASE=test;UID=test;PWD=Test@123"
-    m_pyodbc.connect.assert_called_once_with(expected_connstr)
